+++ kompendium-locations/src/main/kotlin/io/bkbn/kompendium/locations/LocationMethodParser.kt
@@ -1,10 +1,12 @@
 package io.bkbn.kompendium.locations
 
+import io.bkbn.kompendium.annotations.Param
 import io.bkbn.kompendium.core.Kompendium
 import io.bkbn.kompendium.core.metadata.method.MethodInfo
 import io.bkbn.kompendium.core.parser.IMethodParser
 import io.bkbn.kompendium.oas.path.Path
 import io.bkbn.kompendium.oas.path.PathOperation
+import io.bkbn.kompendium.oas.payload.Parameter
 import io.ktor.application.feature
 import io.ktor.locations.KtorExperimentalLocationsAPI
 import io.ktor.locations.Location
@@ -12,23 +14,37 @@ import io.ktor.routing.Route
 import io.ktor.routing.application
 import kotlin.reflect.KAnnotatedElement
 import kotlin.reflect.KClass
+import kotlin.reflect.KClassifier
 import kotlin.reflect.KType
+import kotlin.reflect.full.createType
 import kotlin.reflect.full.findAnnotation
 import kotlin.reflect.full.hasAnnotation
 import kotlin.reflect.full.memberProperties
 
 @OptIn(KtorExperimentalLocationsAPI::class)
 object LocationMethodParser : IMethodParser {
-//  override fun KType.toParameterSpec(info: MethodInfo<*, *>, feature: Kompendium): List<Parameter> {
-//    TODO()
-//  }
+  override fun KType.toParameterSpec(info: MethodInfo<*, *>, feature: Kompendium): List<Parameter> {
+    val clazzList = determineLocationParents(classifier!!)
+    return clazzList.associateWith { it.memberProperties }
+      .flatMap { (clazz, memberProperties) -> memberProperties.associateWith { clazz }.toList() }
+      .filter { (prop, _) -> prop.hasAnnotation<Param>() }
+      .map { (prop, clazz) -> prop.toParameter(info, clazz.createType(), clazz, feature) }
+  }
 
-  @Suppress("UnusedPrivateMember")
-  private inline fun <reified T> yoinkParent() {
-    val clazz = T::class as KClass<*>
+  private fun determineLocationParents(classifier: KClassifier): List<KClass<*>> {
+    var clazz: KClass<*>? = classifier as KClass<*>
+    val clazzList = mutableListOf<KClass<*>>()
+    while (clazz != null) {
+      clazzList.add(clazz)
+      clazz = getLocationParent(clazz)
+    }
+    return clazzList
+  }
+
+  private fun getLocationParent(clazz: KClass<*>): KClass<*>? {
     val parent = clazz.memberProperties
       .find { (it.returnType.classifier as KAnnotatedElement).hasAnnotation<Location>() }
-    println(parent?.returnType)
+    return parent?.returnType?.classifier as? KClass<*>
   }
 
   fun KClass<*>.calculateLocationPath(suffix: String = ""): String {